requirements/prod.txt
# Install everything that needs to be compiled
-r compiled.txt

# Django stuff
Django>=1.4,<1.5

# Templates
Markdown

# Logging
raven

# Celery: Message queue
#celery
#django-celery

# L10n/i18n
#Babel
#-e git://github.com/clouserw/tower.git#egg=tower

# Caching
python-memcached

# Configurations
django-environ

# Database backup
django-dbbackup==1.80.1
django-model-utils==2.2

# Extensions
django-extensions

# Forms
django-crispy-forms
django-parsley
django-simple-captcha

# Migrations
South

# Sass/Compass
django-compressor
compressor-compass

#API
djangorestframework==3.1.3
drf-extensions==0.2.2
djangorestframework-jwt==1.7.2
djangorestframework-filters==0.2.2
